AI_Assesment/requirements.txt
# openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
google-generativeai>=0.3.0
openai>=1.0.0
azure-identity>=1.15.0